LDAP based authentication handler for Leihs
This is a fully functional authenticator for Leihs with an LDAP backend.
Threat it as a technology demo. The current state is not fit to be run
in production. Most of the error handling is still missing and users
will get nasty error pages if something goes wrong (e.g. they mistype
their password).