Add Fallback Email
If a user registers using, for example, the username `testuser`, this
will be sent to the authenticator as `email` field of the token. This
will then later cause trouble if we send back this obviously invalid
email address when registering the user and when sending the success
token.

This patch allows the authenticator to use an email address provided by
LDAP instead, discarding invalid email addresses along the way.

This behavior is configurable. You can:

- Disable the feature
- Only use this to replace invalid email addresses
- Always use the email address from LDAP

This fixes #18