Awitch to API Token
This patch switches to using API tokens for communication with the Leihs
API instead of logging in as a user and keeping that session for every
request. This simplifies the code quite a bit and also makes
communication faster since less requests are now necessary to
communicate with Leihs.

This patch changes the mandatory configuration. Adopters now have to
provide an API token but no longer have to provide user credentials.

This fixes #22