Properly Apply GPL
This patch properly applies the GPL to this project, adding license
headers to all source files and adding a GitHub Actions workflow to
check for files with missing headers.

The check is far from being perfect, but should at least never cause
false positives and should make reasonably sure that there is a header
somewhere.