Avoid Implicit String Conversion
This patch stores all attribute data retrieved from LDAP in a dictionary
with all attribute values being in the form of a list, instead of
sometimes being a list and sometimes being a value. This also means that
we don't have to rely on converting the values to strings working fine
which hoped to always work before.