Raise No Error on Auth System Conflict
This patch causes the registration of the authentication system to
handle a conflict status code as success since the authentication system
seems to be already configured.

This makes the creation of this resource now work similar to the
creation of other resources like users or groups where we also accept a
request as successful if the resource we tried to create already
existed.